requirements.txt
geopy>=2.4.0
facenet-pytorch>=2.6.0
scikit-learn>=1.8.0
Pillow>=10.2.0
numpy>=1.26.0
open-clip-torch>=3.3.0
openai-whisper>=20250625
